requirements.txt
pandas==1.1.3
flask==1.1.2
textslack==0.1.5
schedule==0.6.0
spotipy==2.16.1
pyodbc==4.0.30
python-dotenv==0.14.0
sqlalchemy==1.3.19
numpy==1.18.4